// v-next/ignition-core/src/types/module.ts
export type BaseArgumentType =
  | number
  | bigint
  | string
  | boolean
  | ContractFuture<string>
  | StaticCallFuture<string, string>
  | EncodeFunctionCallFuture<string, string>
  | ReadEventArgumentFuture
  | RuntimeValue;

/**
 * Argument type that smart contracts can receive in their constructors and functions.
 *
 * @beta
 */
export type ArgumentType =
  | BaseArgumentType
  | ArgumentType[]
  | { [field: string]: ArgumentType };
// ...
export type ContractFuture<ContractNameT extends string> =
  | NamedArtifactContractDeploymentFuture<ContractNameT>
  | ContractDeploymentFuture
  | NamedArtifactLibraryDeploymentFuture<ContractNameT>
  | LibraryDeploymentFuture
  | NamedArtifactContractAtFuture<ContractNameT>
  | ContractAtFuture;
// ...
export interface StaticCallFuture<
  ContractNameT extends string,
  FunctionNameT extends string,
> {
  type: FutureType.STATIC_CALL;
  id: string;
  module: IgnitionModule;
  dependencies: Set<Future | IgnitionModule>;
  contract: ContractFuture<ContractNameT>;
  functionName: FunctionNameT;
  nameOrIndex: string | number;
  args: ArgumentType[];
  from: string | AccountRuntimeValue | undefined;
}

/**
 * A future representing the encoding of a contract function call
 *
 * @beta
 */
export interface EncodeFunctionCallFuture<
  ContractNameT extends string,
  FunctionNameT extends string,
> {
  type: FutureType.ENCODE_FUNCTION_CALL;
  id: string;
  module: IgnitionModule;
  dependencies: Set<Future | IgnitionModule>;
  contract: ContractFuture<ContractNameT>;
  functionName: FunctionNameT;
  args: ArgumentType[];
}
// ...
export interface ReadEventArgumentFuture {
  type: FutureType.READ_EVENT_ARGUMENT;
  id: string;
  module: IgnitionModule;
  dependencies: Set<Future | IgnitionModule>;
  futureToReadFrom:
    | NamedArtifactContractDeploymentFuture<string>
    | ContractDeploymentFuture
    | SendDataFuture
    | ContractCallFuture<string, string>;
  eventName: string;
  nameOrIndex: string | number;
  emitter: ContractFuture<string>;
  eventIndex: number;
}
// ...
export type BaseSolidityParameterType = number | bigint | string | boolean;

/**
 * Types that can be passed across the Solidity ABI boundary.
 *
 * @beta
 */
export type SolidityParameterType =
  | BaseSolidityParameterType
  | SolidityParameterType[]
  | { [field: string]: SolidityParameterType };

/**
 * Type of module parameters's values.
 *
 * @beta
 */
export type ModuleParameterType = SolidityParameterType | AccountRuntimeValue;

/**
 * The different runtime values supported by Ignition.
 *
 * @beta
 */
export enum RuntimeValueType {
  ACCOUNT = "ACCOUNT",
  MODULE_PARAMETER = "MODULE_PARAMETER",
}

/**
 * A value that's only available during deployment.
 *
 * @beta
 */
export type RuntimeValue =
  | AccountRuntimeValue
  | ModuleParameterRuntimeValue<ModuleParameterType>;